// web/src/generated/addresses.ts
export const ADDRESS_KEYS = {
  ROLE_MANAGER: "0x280d42ce2d22f9db3eaf7241fbb343003df19c139d5010599d4fda142ae25587",
  PROTOCOL_TIMELOCK: "0xdedad1102523537f8c76cd3469e02db27b77c16d9f8dd19cbfe61b73c5d3ba31",
  ORGANIZATION_REGISTRY: "0xc891cebff5415cb200121eb06b40b517e00c6ff834dd68c51afeb897485fe426",
  CREDENTIAL_REGISTRY: "0xb958d0b226cb5611a3c26613e6f334245d98e3dea6d7a3142e1202882b563113",
  ASSET_REGISTRY: "0x3e28a8ad3827d052d7de332cd5888b85aaa54ddce49ab13f9ba8e39bb38486e8",
  ASSET_OWNERSHIP: "0x461475426bdc03352c334c1a738bdb507f1fd631ab44904a1651ef41ee771b42",
  AIRCRAFT_REGISTRY: "0x66795e0c47c93e1c6b8c5c6ef6e0cbf8568f6ec94245aa478f697b0b9b0e4e07",
  COMPONENT_REGISTRY: "0x2aafd1ac3274cb6f14046781ef246be6bb82f7edc2eab5f9e45ad5b029bc0e61",
  DOCUMENT_REGISTRY: "0x127efdffa11c90f5a6ad3ea89e4b9a0f2e501acccf08c745a7db6be53f1600fa",
  MAINTENANCE_REGISTRY: "0xbbbc284dc95c3a841dc5dc8b052ea5d5550db66938332c797e85dbf52f1b6923",
  ASSET_PASSPORT: "0x743f58de56551e47c60fc8a08aa2966ce38a5ec1061a7ad799e2e090ea7be607",
  MARKETPLACE: "0x6c23addfb32f30983cfe7c36304ac8a5fa1fc0bc1cbc451e9c0ebfabdc2f2299",
  FEE_MANAGER: "0x21624c61f7f124980da62249a4eb2d4c29d77a091f2eeff1302d4835fcccaecd",
  ESCROW_FACTORY: "0xb75f9f78c4d94eeeba739a7668b80666d96838d2ccead38843d1136fd0fc3925",
}
// ...
export type AddressBook = Record<keyof typeof ADDRESS_KEYS, Address>;